Split diplocephalus into 2
