+++ config/splat.us.stno0.yaml
@@ -78,6 +78,7 @@ segments:
       - [0x414A8, .rodata, 4E2E0]
       - [0x414B8, .rodata, e_stage_name]
       - [0x414EC, .rodata, e_diplocephalus]
+      - [0x41540, .rodata, e_diplocephalus_2]
       - [0x415B8, .rodata, e_skelerang]
       - [0x41624, .rodata, e_plate_lord]
       - [0x416A0, .rodata, e_plate_lord_2]
@@ -112,6 +113,7 @@ segments:
       - [0x4E2E0, c]
       - [0x4E654, c, e_stage_name]
       - [0x4F4A8, c, e_diplocephalus]
+      - [0x50990, c, e_diplocephalus_2]
       - [0x5191C, c, e_skelerang]
       - [0x52374, c, e_plate_lord]
       - [0x5542C, c, e_plate_lord_2]
